#48 Refactor : Wine 관련 DTO 이름 수정 및 패키징
