--- src/main/java/com/drinkeg/drinkeg/dto/TastingNoteDTO/request/NoteWineRequestDTO.java
@@ -1,16 +1,0 @@
-package com.drinkeg.drinkeg.dto.TastingNoteDTO.request;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class NoteWineRequestDTO {
-
-    private String wineName;
-
-}
