+++ src/main/java/com/drinkeg/drinkeg/service/wineService/WineService.java
@@ -3,14 +3,14 @@ package com.drinkeg.drinkeg.service.wineService;
 import com.drinkeg.drinkeg.domain.Member;
 import com.drinkeg.drinkeg.domain.Wine;
 import com.drinkeg.drinkeg.dto.HomeDTO.RecommendWineDTO;
-import com.drinkeg.drinkeg.dto.TastingNoteDTO.request.NoteWineRequestDTO;
-import com.drinkeg.drinkeg.dto.WineDTO.SearchWineResponseDTO;
+import com.drinkeg.drinkeg.dto.WineDTO.request.SearchWineRequestDTO;
+import com.drinkeg.drinkeg.dto.WineDTO.response.SearchWineResponseDTO;
 
 import java.util.List;
 
 public interface WineService {
 
-    public List<SearchWineResponseDTO> searchWinesByName(NoteWineRequestDTO noteWineRequestDTO);
+    public List<SearchWineResponseDTO> searchWinesByName(SearchWineRequestDTO searchWineRequestDTO);
 
     public Wine findWineById(Long wineId);
