+++ src/main/java/com/drinkeg/drinkeg/service/wineService/WineServiceImpl.java
@@ -5,8 +5,8 @@ import com.drinkeg.drinkeg.converter.WineConverter;
 import com.drinkeg.drinkeg.domain.Member;
 import com.drinkeg.drinkeg.domain.Wine;
 import com.drinkeg.drinkeg.dto.HomeDTO.RecommendWineDTO;
-import com.drinkeg.drinkeg.dto.TastingNoteDTO.request.NoteWineRequestDTO;
-import com.drinkeg.drinkeg.dto.WineDTO.SearchWineResponseDTO;
+import com.drinkeg.drinkeg.dto.WineDTO.request.SearchWineRequestDTO;
+import com.drinkeg.drinkeg.dto.WineDTO.response.SearchWineResponseDTO;
 import com.drinkeg.drinkeg.exception.GeneralException;
 import com.drinkeg.drinkeg.repository.WineRepository;
 import lombok.RequiredArgsConstructor;
@@ -24,10 +24,10 @@ public class WineServiceImpl implements WineService {
     private final WineRepository wineRepository;
 
     @Override
-    public List<SearchWineResponseDTO> searchWinesByName(NoteWineRequestDTO noteWineRequestDTO) {
+    public List<SearchWineResponseDTO> searchWinesByName(SearchWineRequestDTO searchWineRequestDTO) {
 
         // 와인 이름으로 와인을 검색한다.
-        String searchName = noteWineRequestDTO.getWineName();
+        String searchName = searchWineRequestDTO.getWineName();
         List<Wine> exactMatchWines = wineRepository.findAllByName(searchName);
 
         // 와인 이름을 공백으로 나누어 검색한다.
